in_http: add support for 'cors_allow_origins' config parameter (#426).
This patch enable a partial and optional CORS support. For clients
that want to talk directly to this source, it allows to set a white
list of Cross Origin clients through the validation of the HTTP
Origin header. Usage:

 <source>
   type http
   port 9880
   cors_allow_origins http://example.com, http://www.treasuredata.com
 </source>

The new configuration key 'cors_allow_origins', allows to specify a
white list of domains, they must be separated by coma.

For those request that do not pass the CORS filter, the service will
respond with a '403 Forbidden' HTTP status code and close the connection.

Note: the CORS validation is based on the HTTP request headers, for hence
is NOT secure.

Signed-off-by: Eduardo Silva <eduardo@monkey.io>