fix to instantiate a section for "init: true, multi: true", when nothing specified
Previously, "init: true, multi: false" sections were not instantiated (just `[]` for instance variable).
But with this behavior, we cannot use default values of these sections, except for the case to set a
blank Config::Element into `conf.elements`.
It looks abnormal, and non-useful feature for plugin authors.

So, this change will improve the situation.