Issue 1726: Fix for in_tcp log corruption under load.
The TCP input plugin shares a single buffer across all connections.  Under load
the buffer sometimes gets truncated.  The suspicion is that concurrent
connections race to modify it.

Long message is received.
Short message is received while the long message is still being parsed.
The short message is parsed so the buffer will be truncated.
The buffer, which now contains "longmessage\nshort\n", is truncated by the
length of the short message.  It is now "longm".
Another message arrives.  Now the buffer is "longmanother\n", which does not
parse.
The whole buffer is thrown away and subsequent messages are received and
handled as usual.
Eventually the pattern repeats.

The fix is to use a per-connection buffer.