output: Prevent flushing threads from consuming too much CPU.
This fixes the CPU spike issue reported by several users.

If one flush thread sets the retry state, this makes other threads
wait until the next retry time. This behaviour itself is fine and ok.

However, in this case the `interval` variable has a meaningless value
because `state.next_time` is behind the system time. And since the
sleep call gets skipped in such a case (output.rb#L1343):

    sleep interval if interval > 0

it effectively makes the waiting threads to consume the resource
aggressively with busy loops.