Fix plugin log time format if format is JSON
The time format for a plugin logger was out of sync with the parent
logger, and the delegation of #time_format and #time_format= made it
impossible to fix the plugin logger's time format without hacks.

This problem was not noticeable if the log format was text, because
time formatting was delegated via #caller_line to the parent logger.
However, in the case of JSON output, the plugin logger's #format_time
method was called, yielding incorrectly-formatted timestamps.

Added tests for both text and JSON, and added tests for format changes
in general while I was in there.

Signed-off-by: Owen Smith <owen@udacity.com>