Fix warning printed when chunk key placeholder not replaced
This used `$1` to print the key name.  However, there are no capture groups in `CHUNK_KEY_PLACEHOLDER_PATTERN`, so `$1` would not be set.  According to the Ruby docs I looked up, `$&` can be used to get the part of the string that matched, which I think will give the correct output in this case.