buffer: Fix calculation of timekey stats
The current approach of using `add_timekey` and `del_timekey` to track
the count of chunks in each timekey doesn't work as intended - the
count is never decreased to zero because of complications in chunk
creation, as a result the `oldest_timekey` stat is almost never updated
after initial flush(es).

This commit provides a simple and not-so-timely alternative, to perform
recalculation in every flush cycle.

Signed-off-by: Ji-Ping Shen <ji-ping.shen@relexsolutions.com>