Support unixtime_micros and unixtime_nanos to inject time
If we collect docker container logs using the fluentd logging
driver, we need to sort the logs by the time to merge logs from
various containers. Although we can now specify unixtime_millis,
logs with line feed sometimes have the same time in milliseconds.
As a result, the logs are sorted in the wrong order.
This commit will resolve the problem by introducing new time types
of unixtime with higher precision.

Signed-off-by: abicky <takeshi.arabiki@gmail.com>